requirements.txt
Flask
yt-dlp
requests
discord
python-dotenv